+++ app-meta.js
@@ -1,9 +1,9 @@
 globalThis.APP_META = Object.freeze({
   name: '拾词',
-  version: '1.1.0',
+  version: '1.1.1',
   author: '落日七号',
-  githubOwner: 'luori7hao',
-  githubRepo: 'shici-memory',
-  repositoryUrl: 'https://github.com/luori7hao/shici-memory',
-  pagesUrl: 'https://luori7hao.github.io/shici-memory/',
+  githubOwner: 'GalaxyOg',
+  githubRepo: 'shici-hub',
+  repositoryUrl: 'https://github.com/GalaxyOg/shici-hub',
+  pagesUrl: 'https://galaxyog.github.io/shici-hub/',
 });